requirements.txt
starlette
python-dotenv
sqlalchemy
fastapi
uvicorn
psycopg2-binary
